python/nexustrade/requirements.txt
# NexusTrade Requirements
# Terminal Bench v2 - Distributed Trading Platform

# ===========================================
# Web Frameworks
# ===========================================
Django>=4.2,<5.0
djangorestframework>=3.14
fastapi>=0.104
uvicorn[standard]>=0.24
starlette>=0.32

# ===========================================
# Database
# ===========================================
psycopg2-binary>=2.9

SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.12

# ===========================================
# Kafka & Messaging
# ===========================================
confluent-kafka>=2.3

aiokafka>=0.9
kafka-python>=2.0

# ===========================================
# Redis & Caching
# ===========================================
redis>=5.0
hiredis>=2.2
aioredis>=2.0

# ===========================================
# Service Discovery & Config
# ===========================================
python-consul>=1.1
# python-consul2>=0.1  

# ===========================================
# Authentication & Security
# ===========================================
PyJWT>=2.8
cryptography>=41.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6

# ===========================================
# Serialization & Validation
# ===========================================
pydantic>=2.5
pydantic-settings>=2.1
marshmallow>=3.20
orjson>=3.9

# ===========================================
# HTTP Clients
# ===========================================
httpx>=0.25
aiohttp>=3.9
requests>=2.31
tenacity>=8.2

# ===========================================
# Observability
# ===========================================
opentelemetry-api>=1.21
opentelemetry-sdk>=1.21
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-django>=0.42b0
opentelemetry-exporter-jaeger>=1.21
prometheus-client>=0.19
structlog>=23.2

# ===========================================
# Async & Concurrency
# ===========================================
celery>=5.3
asyncio>=3.4
aio-pika>=9.3

# ===========================================
# Utilities
# ===========================================
python-dateutil>=2.8
pytz>=2023.3
arrow>=1.3
pendulum>=3.0
python-dotenv>=1.0
click>=8.1
rich>=13.7

# ===========================================
# Testing
# ===========================================
pytest>=7.4
pytest-django>=4.7
pytest-asyncio>=0.21
pytest-cov>=4.1
pytest-xdist>=3.5
pytest-timeout>=2.2
pytest-mock>=3.12
pytest-randomly>=3.15
factory-boy>=3.3
faker>=22.0
hypothesis>=6.92
responses>=0.24
aioresponses>=0.7
freezegun>=1.2
testcontainers>=3.7

# ===========================================
# Chaos Engineering
# ===========================================
toxiproxy-python>=0.1
# chaos-monkey>=0.1  

# ===========================================
# Development
# ===========================================
ipython>=8.18
black>=23.12
isort>=5.13
mypy>=1.7
ruff>=0.1
